+++ lib/src/code_generators/v3/swagger_requests_generator_v3.dart
@@ -171,6 +171,10 @@ $allMethodsContent
               neededResponse.schema.items?.ref?.split('/')?.last);
     }
 
+    if (neededResponse.schema?.ref != null) {
+      return neededResponse.schema?.ref?.split('/')?.last;
+    }
+
     if (neededResponse.schema?.originalRef != null) {
       return neededResponse.schema.originalRef;
     }
@@ -320,6 +324,8 @@ abstract class $className extends ChopperService''';
       parameterType = parameter.name.capitalize;
     } else if (parameter.schema?.originalRef != null) {
       parameterType = parameter.schema.originalRef;
+    } else if (parameter.schema?.ref != null) {
+      parameterType = parameter.schema.ref.split('/').last;
     } else {
       parameterType = defaultBodyParameter;
     }