+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.0+2
+
+* Fix score on pub.dev
+
+## 1.0.0+1
+
+* Fixed bug with request parameter and result type, when schema.ref used
+* Added real example of usage with Pet Store
+
 ## 1.0.0+1
 
 * Add more info to readme